% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{aggregateFieldCombiner_sqlserver}
\alias{aggregateFieldCombiner_sqlserver}
\title{将聚合字段列表}
\usage{
aggregateFieldCombiner_sqlserver(fieldName_vec = NULL,
  fieldName_caption = NULL, aggr_fun_vec = NULL)
}
\arguments{
\item{fieldName_vec}{字段列表}

\item{fieldName_caption}{字段标题}

\item{agg_fun_vec}{聚合函数}
}
\value{
返回值
}
\description{
将聚合字段列表
}
\examples{
aggregateFieldCombiner();
aggregateFieldCombiner(c('aa','bb'),c('物料数量','物料金额'),c('sum','sum'));
aggregateFieldCombiner(c('aa','bb'),aggr_fun_vec = c('sum','sum'));
}
