% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-dbms.R
\name{conn,sqlserver_jdbc_driver-method}
\alias{conn,sqlserver_jdbc_driver-method}
\title{用于生成conn相应的sqlserver_jdbc_driver对应的连接信息}
\usage{
\S4method{conn}{sqlserver_jdbc_driver}(jdbc_driver, server_ip, port = "1433", db_name, username, password)
}
\arguments{
\item{jdbc_driver}{sqlserver_jdbc_driver.}
}
\value{
返回值
}
\description{
用于生成conn相应的sqlserver_jdbc_driver对应的连接信息
}
\examples{
conn();
}
