% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-dbms.R
\name{conn}
\alias{conn}
\title{定义连接的通用函数}
\usage{
conn(jdbc_driver, server_ip, port, db_name, username, password)
}
\arguments{
\item{server_ip}{服务器地址}

\item{db_name}{数据库名称}

\item{username}{用户名}

\item{password}{密码}

\item{x}{名称对}
}
\value{
返回值
}
\description{
定义连接的通用函数
}
\examples{
conn();
}
