% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{conn_config}
\alias{conn_config}
\title{创建连接信息}
\usage{
conn_config(config_file = "data-raw/conn_config_rds.R")
}
\arguments{
\item{config_file}{配置文件}
}
\value{
返回值
}
\description{
创建连接信息
}
\examples{
conn_static_info()
}
