% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-aaaparam.R
\name{conn_demo_setting}
\alias{conn_demo_setting}
\title{进入默认演示服务器}
\usage{
conn_demo_setting(db_name = "AIS20190427230019")
}
\arguments{
\item{db_name}{数据库名称}
}
\value{
返回值
}
\description{
进入默认演示服务器
}
\examples{
conn_demo_setting();
}
