% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-dbms.R
\name{db_dropTable}
\alias{db_dropTable}
\title{删除数据库中的表结构}
\usage{
db_dropTable(conn, table_name)
}
\arguments{
\item{conn}{连接信息}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
删除数据库中的表结构
}
\examples{
db_dropTable()
}
