% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{db_getColName}
\alias{db_getColName}
\title{获取字段列名}
\usage{
db_getColName(conn = tsda::conn_rds("jlrds"), table_name = "t_mrpt_division")
}
\arguments{
\item{conn}{连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
获取字段列名
}
\examples{
db_getColName()
}
