% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-dbms.R
\name{db_is_newTable}
\alias{db_is_newTable}
\title{判断是否为新表}
\usage{
db_is_newTable(conn, table_name)
}
\arguments{
\item{conn}{数据库连接}

\item{table_name}{数据库表名}
}
\value{
返回值
}
\description{
判断是否为新表
}
\examples{
db_is_newTable();
}
