% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-dbms.R
\name{db_readTable}
\alias{db_readTable}
\title{读取整表数据}
\usage{
db_readTable(conn, table_name)
}
\arguments{
\item{conn}{数据库连接}

\item{table_name}{表名}
}
\value{
返回值
}
\description{
读取整表数据
}
\examples{
db_readTable();
}
