% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{deal_res_type}
\alias{deal_res_type}
\title{处理返回类型问题}
\usage{
deal_res_type(x, res_type = "vector")
}
\arguments{
\item{x}{原始数据}

\item{res_type}{返回值类型}
}
\value{
返回值
}
\description{
处理返回类型问题
}
\examples{
deal_res_type();
}
