% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030-excel.R
\name{excel_getSheetNames}
\alias{excel_getSheetNames}
\title{读取excel页答名称}
\usage{
excel_getSheetNames(file)
}
\arguments{
\item{file}{文件}
}
\value{
返回值
}
\description{
读取excel页答名称
}
\examples{
execl_getSheetNames()
}
