% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{groupByCombiner_sqlserver}
\alias{groupByCombiner_sqlserver}
\title{确定分组字段}
\usage{
groupByCombiner_sqlserver(fieldName_vec = NULL)
}
\arguments{
\item{fieldName_vec}{分组字段列表}
}
\value{
返回值
}
\description{
确定分组字段
}
\examples{
groupByCombiner_sqlserver();
groupByCombiner_sqlserver(LETTERS);
}
