% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-DBMS.R
\docType{class}
\name{jdbc_driver-class}
\alias{jdbc_driver-class}
\title{定义一下虚拟的jdbc_dirver类}
\value{
返回值
}
\description{
定义一下虚拟的jdbc_dirver类
}
\section{Slots}{

\describe{
\item{\code{driver_name}}{character. 驱动类型}

\item{\code{driver_location}}{character.驱动位置}

\item{\code{os_type}}{character. 操作系统类型}

\item{\code{db_type}}{数据库类型}
}}

\examples{
jdbc_dirver
}
