% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{logical_as_ascs}
\alias{logical_as_ascs}
\title{排序判断辅助函数}
\usage{
logical_as_ascs(logical_vec)
}
\arguments{
\item{logical_vec}{逻辑变量}
}
\value{
返回值
}
\description{
排序判断辅助函数
}
\examples{
logical_as_ascs();
}
