% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/043-mysql.R
\name{mysql_conn_common}
\alias{mysql_conn_common}
\title{mysql 数据库连接设置}
\usage{
mysql_conn_common(
  ip = "localhost",
  port = 3306,
  user_name = "root",
  password = "Hoolilay889@",
  db_name = "rdkc"
)
}
\arguments{
\item{ip}{服务器IP地址}

\item{port}{商品号}

\item{user_name}{用户名}

\item{password}{密码}

\item{db_name}{数据库名称}
}
\value{
返回值
}
\description{
mysql 数据库连接设置
}
\examples{
mysql_conn_common();
}
