% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_answ_template_count}
\alias{nsim_answ_template_count}
\title{处理答案的标签个数}
\usage{
nsim_answ_template_count()
}
\value{
返回值
}
\description{
处理答案的标签个数
}
\examples{
nsim_answ_template_count()
}
