% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_brand_fieldNames}
\alias{nsim_brand_fieldNames}
\title{处理品牌字段信息}
\usage{
nsim_brand_fieldNames(lang = "en", res_type = "vector")
}
\arguments{
\item{lang}{返回语言格式}

\item{res_type}{返回字段格式}
}
\value{
返回值
}
\description{
处理品牌字段信息
}
\examples{
nsim_brand_fieldNames()
}
