% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_brand_FName}
\alias{nsim_brand_FName}
\title{获取字段名}
\usage{
nsim_brand_FName(res_type = "vector")
}
\arguments{
\item{res_type}{返回字段类型}
}
\value{
返回值
}
\description{
获取字段名
}
\examples{
nsim_brand_FName();
nsim_brand_FName();
nsim_brand_FName('list');
}
