% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_brand_FNumber}
\alias{nsim_brand_FNumber}
\title{读取品牌中的数据类型}
\usage{
nsim_brand_FNumber(res_type = "vector")
}
\arguments{
\item{res_type}{返回值类型}
}
\value{
返回值
}
\description{
读取品牌中的数据类型
}
\examples{
nsim_brand_FNumber();
nsim_brand_FNumber('list');
}
