% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_brand_table_query}
\alias{nsim_brand_table_query}
\title{增加通用的品牌信息查询}
\usage{
nsim_brand_table_query(table_name = "qalist_raw",
  field_name = "FQuestionText", brand = "JBLH", unique = FALSE)
}
\arguments{
\item{table_name}{表名}

\item{field_name}{字段名}

\item{brand}{品牌}

\item{unique}{是否唯一}
}
\value{
返回值
}
\description{
增加通用的品牌信息查询
}
\examples{
nsim_brand_table_query();
}
