% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_brand_table_query2}
\alias{nsim_brand_table_query2}
\title{处理表查询的问题}
\usage{
nsim_brand_table_query2(
  table_name = "qalist_raw",
  field_name = "FQuestionText",
  brand = "JBLH",
  unique = FALSE
)
}
\arguments{
\item{table_name}{表名称}

\item{field_name}{字段名}

\item{brand}{品牌}

\item{unique}{是否唯一}
}
\value{
返回值
}
\description{
处理表查询的问题
}
\examples{
nsim_brand_table_query2();
}
