% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40A-sqlserver_nsim.R
\name{nsim_fieldNames}
\alias{nsim_fieldNames}
\title{获取表的字段名向量}
\usage{
nsim_fieldNames(table_name = "test2", id_var = NULL,
  res_type = "vector")
}
\arguments{
\item{table_name}{表名}

\item{id_var}{内码字段名}

\item{res_type}{返回字段类型vector or sql}
}
\value{
返回值
}
\description{
获取表的字段名向量
}
\examples{
nsim_filedNames();
}
