% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_im_raw_record_fieldNames}
\alias{nsim_im_raw_record_fieldNames}
\title{处理语料的字段信息}
\usage{
nsim_im_raw_record_fieldNames(lang = "en", res_type = "vector")
}
\arguments{
\item{lang}{语言选择}

\item{res_type}{返回类型选择}
}
\value{
返回值
}
\description{
处理语料的字段信息
}
\examples{
nsim_im_raw_record_fieldNames();
}
