% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_im_raw_record_FLog}
\alias{nsim_im_raw_record_FLog}
\title{处理原始语料的日志字段}
\usage{
nsim_im_raw_record_FLog(res_type = "vector")
}
\arguments{
\item{res_type}{返回值类型}
}
\value{
返回值
}
\description{
处理原始语料的日志字段
}
\examples{
nsim_im_raw_record_FLog();
}
