% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/40A-sqlserver_nsim.R
\name{nsim_max_id}
\alias{nsim_max_id}
\title{读取表的内码最大值}
\usage{
nsim_max_id(table_name = "test2", id_var = "FInterId")
}
\arguments{
\item{table_name}{表名}

\item{id_var}{内码字段名}
}
\value{
返回值
}
\description{
读取表的内码最大值
}
\examples{
nsim_max_id();
}
