% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_qalist_FAnswerText}
\alias{nsim_qalist_FAnswerText}
\title{处理QA中的answer问题}
\usage{
nsim_qalist_FAnswerText(brand = "JBLH", unique = FALSE)
}
\arguments{
\item{brand}{品牌}

\item{unique}{唯一性}
}
\value{
返回值
}
\description{
处理QA中的answer问题
}
\examples{
nsim_qalist_FAnswerText();
}
