% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_qalist_FAnswerText2}
\alias{nsim_qalist_FAnswerText2}
\title{处理回复字段2}
\usage{
nsim_qalist_FAnswerText2(brand = "JBLH", unique = FALSE)
}
\arguments{
\item{brand}{品牌}

\item{unique}{是否回复}
}
\value{
返回值
}
\description{
处理回复字段2
}
\examples{
nsim_qalist_FAnswerText2();
}
