% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_qalist_FQuestionText}
\alias{nsim_qalist_FQuestionText}
\title{提取品牌的文本信息}
\usage{
nsim_qalist_FQuestionText(brand = "JBLH", unique = FALSE)
}
\arguments{
\item{brand}{品牌}

\item{unique}{是否唯一}
}
\value{
返回值
}
\description{
提取品牌的文本信息
}
\examples{
nsim_qalist_FQuestionText();
}
