% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_ques_template_count}
\alias{nsim_ques_template_count}
\title{读取问题模板的计数数据}
\usage{
nsim_ques_template_count()
}
\value{
返回值
}
\description{
读取问题模板的计数数据
}
\examples{
nsim_ques_template_count();
}
