% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_read_byField}
\alias{nsim_read_byField}
\title{读取表中的字段}
\usage{
nsim_read_byField(
  table_name = "test2",
  field_name = "FName",
  res_type = "vector"
)
}
\arguments{
\item{table_name}{表名}

\item{field_name}{字段名}

\item{res_type}{返回值类型}
}
\value{
返回值
}
\description{
读取表中的字段
}
\examples{
nsim_read_byField();
}
