% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_read_distinct}
\alias{nsim_read_distinct}
\title{处理查询的唯一性问题}
\usage{
nsim_read_distinct(table_name = "test2", id_var = NULL,
  field_vars = NULL, where = " fid =1", unique = FALSE)
}
\arguments{
\item{table_name}{表名}

\item{id_var}{内码}

\item{field_vars}{名称}

\item{where}{查询条件}

\item{unique}{是否已到}
}
\value{
返回值
}
\description{
处理查询的唯一性问题
}
\examples{
nsim_read_distinct();
}
