% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_read_where}
\alias{nsim_read_where}
\title{处理语句读取问题}
\usage{
nsim_read_where(table_name = "test2", id_var = NULL,
  field_vars = NULL, where = " fid =1")
}
\arguments{
\item{table_name}{表名}

\item{id_var}{内码}

\item{field_vars}{字段名}

\item{where}{条件}
}
\value{
返回值
}
\description{
处理语句读取问题
}
\examples{
nsim_read_where();
nsim_read_where('qalist_raw',field_vars = 'FQuestionText',where = "FBrand = 'JBLH'");
}
