% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_save}
\alias{nsim_save}
\title{将相同结构的数据写入到数据库中}
\usage{
nsim_save(data, table_name = "test2", id_var = NULL)
}
\arguments{
\item{data}{数据}

\item{table_name}{写入的表名}
}
\value{
返回值
}
\description{
将相同结构的数据写入到数据库中
}
\examples{
nsim_save();
}
