% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_split_combine}
\alias{nsim_split_combine}
\title{处理数据按||合并的问题}
\usage{
nsim_split_combine(x, group)
}
\arguments{
\item{x}{向量}

\item{group}{分量}
}
\value{
返回数据框
}
\description{
处理数据按||合并的问题
}
\examples{
nsim_split_combine();
}
