% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_version_create}
\alias{nsim_version_create}
\title{创建新的版本}
\usage{
nsim_version_create(brand = "JBLH", module = "nsbl", from = 1L,
  to = 1000L, prefix = "V")
}
\arguments{
\item{brand}{品牌}

\item{module}{模块}

\item{from}{开始}

\item{to}{结束}

\item{prefix}{前缀}
}
\value{
返顺值
}
\description{
创建新的版本
}
\examples{
nsim_version_create();
nsim_version_create('nsdict');
nsim_version_create('nsbl',1L,10000L);
}
