% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_version_getCurrentVersion}
\alias{nsim_version_getCurrentVersion}
\title{获取品牌模块的版本}
\usage{
nsim_version_getCurrentVersion(brand = "JBLH", module = "nsbl")
}
\arguments{
\item{brand}{品牌}

\item{module}{模块}
}
\value{
返回值
}
\description{
获取品牌模块的版本
}
\examples{
nsim_version_getCurrentVersion();
}
