% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{nsim_version_setCurrentVersion}
\alias{nsim_version_setCurrentVersion}
\title{设置品牌模块的版本}
\usage{
nsim_version_setCurrentVersion(brand = "JBLH", module = "nsbl", version)
}
\arguments{
\item{brand}{品牌}

\item{module}{模块}

\item{version}{版本}
}
\value{
无
}
\description{
设置品牌模块的版本
}
\examples{
nsim_version_setCurrentVersion();
}
