% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{orderByCombiner_sqlserver}
\alias{orderByCombiner_sqlserver}
\title{处理order By子句}
\usage{
orderByCombiner_sqlserver(fieldName_vec = NULL, asc_vec = NULL)
}
\arguments{
\item{fieldName_vec}{字段列表}

\item{asc_vec}{排序向量}
}
\value{
返回值
}
\description{
处理order By子句
}
\examples{
orderByCombiner();
orderByCombiner();
orderByCombiner(fieldName_vec = letters,asc_vec = rep(T,26));
}
