% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-directory.R
\name{print.dir_files}
\alias{print.dir_files}
\title{打印dir_files文件}
\usage{
\method{print}{dir_files}(object)
}
\arguments{
\item{object}{dir_files类型的对象}
}
\value{
返回值
}
\description{
打印dir_files文件
}
\examples{
print();
}
