% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010-directory.R
\name{print_dir}
\alias{print_dir}
\title{显示目录下所有文件并打印出来，每个文件换行}
\usage{
print_dir(dir_path)
}
\arguments{
\item{dir_path}{目录路径}
}
\value{
返回值
}
\description{
显示目录下所有文件并打印出来，每个文件换行
}
\examples{
print_dir();
}
