% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030-excel.R
\name{read_excel_GH}
\alias{read_excel_GH}
\title{广汇数据处理后}
\usage{
read_excel_GH(gh_download_dir = "./data-raw/test")
}
\arguments{
\item{gh_download_dir}{文件目录}
}
\value{
返回值
}
\description{
广汇数据处理后
}
\examples{
read_excel_GH();
}
