% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030-excel.R
\name{read_excelBooks}
\alias{read_excelBooks}
\title{将多个相同结构的EXCEL合并出一个数据}
\usage{
read_excelBooks(dir_files)
}
\arguments{
\item{dir_files}{excel_files所在的目录}
}
\value{
返回值
}
\description{
将多个相同结构的EXCEL合并出一个数据
}
\examples{
read_excelBooks('./data-raw/ghdata')
}
