% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030-excel.R
\name{read_excelSheet}
\alias{read_excelSheet}
\title{读取excel的默认第一个页签}
\usage{
read_excelSheet(excel_file)
}
\arguments{
\item{excel_file}{文件名}
}
\value{
返回值
}
\description{
读取excel的默认第一个页签
}
\examples{
read_excelSheet('./data-raw/gh_llb.xlsx')
}
