% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData.excel}
\alias{readData.excel}
\title{读取Excel文件}
\usage{
\method{readData}{excel}(file)
}
\arguments{
\item{file}{文件名，class=excel.}
}
\value{
data.frame.
}
\description{
读取Excel文件
}
\examples{
readData(fileExcel);
}
