% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData}
\alias{readData}
\title{读取数据的通知格式函数}
\usage{
readData(file = "", format = "excel", pkg = "readxl")
}
\arguments{
\item{file}{输入文件名}

\item{foramt}{选择excel等}
}
\value{
返回结果
}
\description{
读取数据的通知格式函数
}
\examples{
readData('./data/test.xls','excel','readxl');
}
