% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData.spss}
\alias{readData.spss}
\title{读取spss数据}
\usage{
\method{readData}{spss}(file)
}
\arguments{
\item{file}{文件为class =spss}
}
\value{
返回数据框
}
\description{
读取spss数据
}
\examples{
readData(fileSpss);
}
