% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{readData.stata}
\alias{readData.stata}
\title{读取stata数据}
\usage{
\method{readData}{stata}(file)
}
\arguments{
\item{file}{数据源为stata}
}
\value{
返回数据框
}
\description{
读取stata数据
}
\examples{
readData(fileStata);
}
