% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030-excel.R
\name{readExcelDf}
\alias{readExcelDf}
\title{读取Excel数据到数据框}
\usage{
readExcelDf(file)
}
\arguments{
\item{file}{数据文件名}
}
\value{
返回值
}
\description{
读取Excel数据到数据框
}
\examples{
readExcelDf('.data-raw/txt.xlsx');
读取excel数据
}
