% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{select_gen_sqlserver}
\alias{select_gen_sqlserver}
\title{定义sqlserver的语句自动生成函数}
\usage{
select_gen_sqlserver(
  table_name = "T_GL_VOUCHER",
  col_vec = c("FDATE", "FYEAR", "FPERIOD", "FBILLNO"),
  row_where_expr = NULL,
  order_vec = NULL,
  order_asc_logical = NULL
)
}
\arguments{
\item{table_name}{表名称}

\item{col_vec}{字段名向量}

\item{row_where_expr}{where表达式，完全使用tsql语法}

\item{order_vec}{排序字段}

\item{order_asc_logical}{逻辑表达式TRUE表示asc,FALSE表示desc}
}
\value{
返回值
}
\description{
定义sqlserver的语句自动生成函数
}
\examples{
bb <- select_gen_sqlserver(table_name='T_GL_VOUCHER',
col_vec=c('FDATE','FYEAR','FPERIOD','FBILLNO'),
row_where_expr="FDATE>='2019-01-01' and fdate <= '2019-03-31' ",order_vec=c('FDATE','FBILLNO'),order_asc_logical=c(T,T));
bb;
cc <- select_gen_sqlserver(table_name='T_GL_VOUCHER',
col_vec=c('FDATE','FYEAR','FPERIOD','FBILLNO'),
row_where_expr=NULL,order_vec=c('FDATE','FBILLNO'),order_asc_logical=c(F,T));
cc;
dd <- select_gen_sqlserver(table_name='T_GL_VOUCHER',
col_vec=c('FDATE','FYEAR','FPERIOD','FBILLNO'),
row_where_expr=NULL);
dd;
}
