% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{simpleFieldCombiner_sqlserver}
\alias{simpleFieldCombiner_sqlserver}
\title{将字段列表进行处理}
\usage{
simpleFieldCombiner_sqlserver(fieldName_vec, fieldName_caption = NULL)
}
\arguments{
\item{fieldName_vec}{字段列表}

\item{fieldName_caption}{字段标题}
}
\value{
返回值
}
\description{
将字段列表进行处理
}
\examples{
simpleFieldCombiner();
simpleFieldCombiner(letters,LETTERS);
simpleFieldCombiner(c('fnumber','fname'),c('物料代码','物料名称'));
}
