% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{sql_conn}
\alias{sql_conn}
\title{添加sql 连接函数}
\usage{
sql_conn(ip, port = 1433, user_name = "sa", password,
  db_name = "test")
}
\arguments{
\item{ip}{服务器地址}

\item{port}{端口}

\item{user_name}{用户名}

\item{password}{密码}

\item{db_name}{数据库}
}
\value{
返回链接
}
\description{
添加sql 连接函数
}
\examples{
sql_conn
}
