% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{sql_conn_rd}
\alias{sql_conn_rd}
\title{读取棱星数据库的配置信息}
\usage{
sql_conn_rd(db_name = "AIS20190427230019")
}
\arguments{
\item{db_name}{数据库名称,其他信息已加密处理}
}
\value{
返回值
}
\description{
读取棱星数据库的配置信息
}
\examples{
sql_conn_rd
}
