% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{sql_fieldInfo}
\alias{sql_fieldInfo}
\title{查询指定表的字段类型}
\usage{
sql_fieldInfo(conn = conn_rds(), table_name = "books")
}
\arguments{
\item{conn}{连接信息}

\item{table_name}{表名}
}
\value{
返回数据库
}
\description{
查询指定表的字段类型
}
\examples{
sql_fieldInfo()
}
