% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{sql_gen_delete}
\alias{sql_gen_delete}
\title{创建sql更新语句}
\usage{
sql_gen_delete(table = "books")
}
\arguments{
\item{table}{表名}
}
\value{
返回值
}
\description{
创建sql更新语句
}
\examples{
sql_gen_delete()
}
