% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{sql_gen_insert}
\alias{sql_gen_insert}
\title{生成insert相关语句}
\usage{
sql_gen_insert(conn = conn_rds("test"), table = "books",
  fieldNames = NULL)
}
\arguments{
\item{conn}{连接}

\item{table}{表名}

\item{fieldNames}{字段名}
}
\value{
返回值
}
\description{
生成insert相关语句
}
\examples{
sql_gen_insert()
}
