% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{sql_gen_select}
\alias{sql_gen_select}
\title{针对列表查询}
\usage{
sql_gen_select(conn = conn_rds("test"), table = "books", fieldNames = NULL)
}
\arguments{
\item{conn}{连接信息}

\item{table}{表名}

\item{fieldNames}{字段清单，向量定义，不建议使用*}
}
\value{
返回值
}
\description{
针对列表查询
}
\examples{
sql_gen_select(table = 't_test')
sql_gen_select(table = 'conn')
sql_gen_select()
}
