% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-DBMS.R
\docType{class}
\name{sqlserver_jdbc_driver-class}
\alias{sqlserver_jdbc_driver-class}
\title{定义sqlserver_jdbc_driver类}
\value{
返回值
}
\description{
定义sqlserver_jdbc_driver类
}
\examples{
sqlserver_jdbc_driver();
}
