% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-dbms.R
\name{sqlserver_jdbc_driver}
\alias{sqlserver_jdbc_driver}
\title{定义sqlserver_jdbc_driver入口实例化函数}
\usage{

  sqlserver_jdbc_driver(driver_location = "/opt/sqljdbc_7.2/enu/mssql-jdbc-7.2.2.jre8.jar",
  os_type = "linux")
}
\arguments{
\item{driver_location}{驱动存放位置}

\item{os_type}{操作系统类型}
}
\value{
返回值
}
\description{
定义sqlserver_jdbc_driver入口实例化函数
}
\examples{
sqlserver_jdbc_driver();
}
