% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{upload_data}
\alias{upload_data}
\title{上传数据到数据，增量部分}
\usage{
upload_data(conn, table_name, data)
}
\arguments{
\item{conn}{数据库连接}

\item{table_name}{表名}

\item{data}{追加数据}
}
\value{
返回值
}
\description{
上传数据到数据，增量部分
}
\examples{
upload_data()
}
