% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040A-sqlserver_nsim.R
\name{vector_as_sql_fieldNames}
\alias{vector_as_sql_fieldNames}
\title{将向量变成SQL字段名}
\usage{
vector_as_sql_fieldNames(x)
}
\arguments{
\item{x}{向量}
}
\value{
字段名
}
\description{
将向量变成SQL字段名
}
\examples{
vector_as_sql_fieldNames();
}
