% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040-sqlserver.R
\name{whereStatementCombiner_sqlserver}
\alias{whereStatementCombiner_sqlserver}
\title{形成一个where子句}
\usage{
whereStatementCombiner_sqlserver(
  fieldName_vec = NULL,
  comparerSig_vec = NULL,
  filterValue_vec = NULL,
  comboCondition_logi_vec = NULL
)
}
\arguments{
\item{fieldName_vec}{字段列表}

\item{comparerSig_vec}{比例符>=<}

\item{filterValue_vec}{比例值}

\item{comboCondition_logi_vec}{and或者or最后一个逻辑不生效}
}
\value{
返回值
}
\description{
形成一个where子句
}
\examples{
 whereStatementCombiner();
is.null(whereStatementCombiner());
fieldName_vec <-c('FYEAR','FDATE');
comparerSig_vec <-c('=','>=');
filterValue_vec <-c("2018","'2018-10-02'");
comboCondition_logi_vec <-c('and','and');
whereStatementCombiner(fieldName_vec,comparerSig_vec,filterValue_vec,comboCondition_logi_vec);
}
