% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{writeData}
\alias{writeData}
\title{将数据写到到某个系统默认为R}
\usage{
writeData(data, to = "R", overwrite = F)
}
\arguments{
\item{data}{系统中R数据对照}

\item{to}{输入的文件系统，默认为R}

\item{overwrite}{是否覆盖原有的数据，默认为否}
}
\value{
不用于返回
}
\description{
将数据写到到某个系统默认为R
}
\examples{
writeData(data,'R',T);
}
