% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeData.R
\name{writeData2Excel}
\alias{writeData2Excel}
\title{将数据写入到Excel}
\usage{
writeData2Excel(data, fileName, sheetName = "sheet1")
}
\arguments{
\item{data}{数据框}

\item{fileName}{目标文件名}

\item{sheetName}{页签名}
}
\value{
返顺值
}
\description{
将数据写入到Excel
}
\examples{
writeData2Excel(letters,'text.xlsx','sheet1');
}
