% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030-excel.R
\name{writeDataToExcel}
\alias{writeDataToExcel}
\title{将R对象写入到Excel文件中}
\usage{
writeDataToExcel(data, fileName, sheetName)
}
\arguments{
\item{data}{R对象}

\item{fileName}{文件名}

\item{sheetName}{页签名}
}
\value{
返回值
}
\description{
将R对象写入到Excel文件中
}
\examples{
writeDataToExcel();
}
